# gateway/api-doc.yaml
openapi: 3.0.0
info:
  title: GreatReads API Gateway
  description: Unified API for GreatReads platform
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:
  /api/users/register:
    post:
      summary: Register a new user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User registered successfully
        "400":
          description: Invalid input

  /api/users/login:
    post:
      summary: Login user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid credentials

  /api/books/search:
    get:
      summary: Buscar libros en Google Books
      tags: [Books]
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Término de búsqueda
      responses:
        "200":
          description: Lista de libros encontrados
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    bookId:
                      type: string
                    title:
                      type: string
                    authors:
                      type: array
                      items:
                        type: string
                    description:
                      type: string
                    publisher:
                      type: string
                    publishedDate:
                      type: string
                    pageCount:
                      type: integer
                    thumbnail:
                      type: string
        "401":
          description: Unauthorized
        "500":
          description: Internal server error

  /api/books/reading:
    post:
      summary: Añadir un libro a la lista de lectura
      tags: [Books]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bookId
              properties:
                bookId:
                  type: string
                progress:
                  type: integer
                  default: 0
                status:
                  type: string
                  example: reading
      responses:
        "201":
          description: Libro añadido correctamente
        "400":
          description: Input inválido
        "401":
          description: Unauthorized
        "500":
          description: Internal server error

    get:
      summary: Obtener la lista de lectura del usuario con info completa
      tags: [Books]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de libros con información completa
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    bookId:
                      type: string
                    title:
                      type: string
                    authors:
                      type: array
                      items:
                        type: string
                    description:
                      type: string
                    publisher:
                      type: string
                    publishedDate:
                      type: string
                    pageCount:
                      type: integer
                    thumbnail:
                      type: string
                    progress:
                      type: integer
                    status:
                      type: string
        "401":
          description: Unauthorized
        "500":
          description: Internal server error

  /api/books/{bookId}:
    get:
      summary: Obtener detalle de un libro por ID
      tags: [Books]
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
          description: ID del libro en Google Books
      responses:
        "200":
          description: Detalle completo del libro
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  authors:
                    type: array
                    items:
                      type: string
                  description:
                    type: string
                  publisher:
                    type: string
                  publishedDate:
                    type: string
                  pageCount:
                    type: integer
                  categories:
                    type: array
                    items:
                      type: string
                  language:
                    type: string
                  thumbnail:
                    type: string
        "500":
          description: Internal server error
          
  /api/books/{bookId}/status:
    get:
      summary: Obtener estado y progreso de un libro para el usuario
      tags: [Books]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
          description: ID del libro en Google Books
      responses:
        "200":
          description: Estado y progreso del libro
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: Leyendo
                  progress:
                    type: integer
                    example: 45
        "401":
          description: Unauthorized
        "500":
          description: Internal server error

  /api/books/reading/{bookId}:
    delete:
      summary: Eliminar un libro de la lista del usuario
      tags: [Books]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Libro eliminado correctamente
        "401":
          description: Unauthorized
        "404":
          description: Libro no encontrado
        "500":
          description: Internal server error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT